sceneRender: fix texture_from_png assumptions
texure_from_png assumed that everything 32 bit must be RGBA, and
everything less than that must be RGB. This is wrong. PNG allows
for transparency even in palette or grayscale mode using the
tRNS chunk.

Let's convert to RGBA if we have transparency, and convert to RGB
if we don't.